fix(debug): unblock PR quality — test imports, split scope tests, no worker phase gate
- Use colonizethis_test import in logic debug_console_workers_test (repo.test_imports).
- Move applyDebugWorkerPoolCredit tests to app_event_handler_debug_worker_pool_test.dart
  to satisfy dart non-comment line size gate on app_event_handler_scope_test.
- Remove TurnPhase.orders gate from applyDebugWorkerPoolCredit (Refs #2389 AC parity).
- Align SPEC/ui/debug-console-panel.md orders policy and ACs for /add_worker.
