+++ SPEC/ui/debug-console-panel.md
@@ -31,7 +31,7 @@
 - **`/get_tile_basic_info`** — read-only command that reports ids from the current orange tile selection (`mapProvincePanelProvider.selectedTileKey`). Success output is multiline and includes `tile_id: <regionId|localProvinceId|x|y>` and `province_id: <regionId|localProvinceId>`. Missing selection returns deterministic error `No tile is selected.`; malformed selection returns deterministic error `Selected tile key is invalid.`.
 - **`/list_players`** — read-only command that enumerates all `Game.players` in ascending `player.id` order. Success output starts with `players_count: <N>` and one blank-line-separated block per player with `player_id`, `display_name` (empty/blank `displayName` falls back to `player_id`), `type` (`human` or `ai`), and `eliminated` (`true` when `capitalProvinceId == null`, else `false`). Extra arguments return `Usage: /list_players`. When the submit-time read-only context does not supply a `players` projection, the executor returns deterministic error `Player list is unavailable.` with no events.
 - **`/help`** — Lists supported commands and bounds. `/spawn_regiment`, `/spawn_ship`, `/add_resource`, and `/add_worker` help text must include every canonical id (regiment, ship, commodity, or worker tier) exactly once in stable sorted order, generated from the same source used for parser validation.
-- **Orders-phase gate policy** — `/add_money`, `/add_worker`, `/add_resource`, `/flip_province`, and `/reveal_province` are allowed only during human Orders phase. Outside Orders phase, the app listener rejects command apply with deterministic feedback and leaves game state unchanged.
+- **Orders-phase gate policy** — `/add_money`, `/add_resource`, `/flip_province`, and `/reveal_province` apply paths are allowed only during human Orders phase. Outside Orders phase, the app listener rejects those applies with deterministic feedback and leaves game state unchanged. **`/add_worker`** is **not** `TurnPhase`-gated in its apply handler (parity with the “no economy-phase modifiers” credit surface for `/add_money` / `/add_worker` above); a valid `CreditDebugWorkerPoolEvent` still mutates `WorkerPool` outside Orders phase.
 
 ---
 
@@ -62,10 +62,10 @@
 - Given panel input `/add_money 500` with the active human player’s treasury at `100`, when the player submits, then the system emits one `CreditDebugTreasuryEvent` with `requestedAmount=500`, `creditedAmount=500`, and after apply the human player’s treasury is `600`.
 - Given panel input `/add_money 12000`, when the player submits, then the system emits one `CreditDebugTreasuryEvent` with `requestedAmount=12000` and `creditedAmount=9999`, and success feedback states both the requested amount (`12000`) and the credited amount (`9999`) plus the resulting treasury balance.
 - Given panel input `/add_money 0` or `/add_money abc`, when the player submits, then the system emits no `CreditDebugTreasuryEvent` and shows a deterministic error message.
-- Given panel input `/add_worker peasants 500` with the active human player’s `WorkerPool.peasants` at `10`, when the player submits during human Orders phase, then the system emits one `CreditDebugWorkerPoolEvent` with `workerTierId=peasants`, `requestedAmount=500`, `creditedAmount=500`, and after apply the human player’s `WorkerPool.peasants` is `510`.
-- Given panel input `/add_worker MASTERS 12000`, when the player submits during human Orders phase, then the parser accepts case-insensitive tier input and emits one `CreditDebugWorkerPoolEvent` with `workerTierId=masters`, `requestedAmount=12000`, and `creditedAmount=9999`, and success feedback states both the requested amount (`12000`) and the credited amount (`9999`) plus the resulting `masters` tier count.
+- Given panel input `/add_worker peasants 500` with the active human player’s `WorkerPool.peasants` at `10`, when the player submits, then the system emits one `CreditDebugWorkerPoolEvent` with `workerTierId=peasants`, `requestedAmount=500`, `creditedAmount=500`, and after apply the human player’s `WorkerPool.peasants` is `510`.
+- Given panel input `/add_worker MASTERS 12000`, when the player submits, then the parser accepts case-insensitive tier input and emits one `CreditDebugWorkerPoolEvent` with `workerTierId=masters`, `requestedAmount=12000`, and `creditedAmount=9999`, and success feedback states both the requested amount (`12000`) and the credited amount (`9999`) plus the resulting `masters` tier count.
 - Given panel input `/add_worker peasants 0`, `/add_worker peasants abc`, or `/add_worker invalid_tier 5`, when the player submits, then the system emits no `CreditDebugWorkerPoolEvent` and shows deterministic validation feedback.
-- Given a valid `/add_worker apprentices 50` command outside human Orders phase, when the app listener attempts apply, then the system keeps game state unchanged and shows deterministic phase-gate rejection feedback.
+- Given a valid `/add_worker apprentices 50` command while the active game’s `TurnState.phase` is **not** `TurnPhase.orders`, when the app listener applies `CreditDebugWorkerPoolEvent`, then the system increases the human player’s `WorkerPool.apprentices` by `50` and persists the mutation (no phase-gate rejection path).
 - Given panel input `/add_resource grain 500` during human Orders phase, when the player submits, then the system emits one `CreditDebugStockpileCommodityEvent` with `commodityId=grain`, `requestedAmount=500`, and `creditedAmount=500`.
 - Given panel input `/add_resource castIron 12000` during human Orders phase, when the player submits, then the system emits one `CreditDebugStockpileCommodityEvent` with `commodityId=castIron`, `requestedAmount=12000`, and `creditedAmount=9999`.
 - Given panel input `/add_resource GRAIN 10` during human Orders phase, when the player submits, then the parser accepts case-insensitive canonical-id input and emits canonical `commodityId=grain`.
